+++ e_epicier/credit/views.py
@@ -3,11 +3,12 @@ from .forms import  CreditForm,SelectedProductForm
 from .models import Credit,Qnt_Produit
 from .models import Produit
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 import decimal
 # Create your views here.
 
-
+@login_required
 def creditAdd(request):
     if request.method == 'POST':
         form = CreditForm(request.POST)
@@ -23,8 +24,10 @@ def creditAdd(request):
         form = CreditForm()
     return render(request,'credits/credit_add.html',{'form':form})
 
+@login_required
 def creditEdit(request,id):
     credit = Credit.objects.get(id=id)
+    if check(credit,request): return redirect('home')
     produits = Qnt_Produit.objects.filter(credit=id)
     if request.method == 'POST' and 'pay' in request.POST:
         creditPay(request,id)
@@ -38,23 +41,25 @@ def creditEdit(request,id):
             return redirect('credit-view')
     else:
         form = CreditForm(instance = credit)
-    return render(request,'credits/credit_edit.html',{'form':form,'produits':produits, 'id':id})
-
+    return render(request,'credits/credit_edit.html',{'form':form,'produits':produits, 'id':id,'credit':credit})
 
 def creditDetail(request,id):
     credits = Credit.objects.get(id=id)
     produits = Qnt_Produit.objects.filter(credit=id)
     return render(request, 'credits/credit_detail.html',{'credits':credits,'produits':produits})
 
-
+@login_required
 def creditDelete(request,id):
     credit = Credit.objects.get(id=id)
+    if check(credit,request): return redirect('home')
     credit.delete()
     return redirect('credit-view')
 
+@login_required
 def creditProduits(request,id):
     produits = Produit.objects.filter(user=request.user).values
     credit = Credit.objects.get(id=id)
+    if check(credit,request): return redirect('home')
     total =0
     if request.method == 'POST':
         for product in Produit.objects.all():
@@ -64,13 +69,15 @@ def creditProduits(request,id):
                 selected_product = Qnt_Produit(produit=product, qnt=quantity, credit = credit,total=quantity*product.prix)
                 selected_product.save()
         credit.to_pay +=total
+        credit.etat = 0
         credit.save()
         return creditView(request)
     return render(request,'produits/produit_credit_add.html',{'produits':produits})
 
-
+@login_required
 def creditPay(request,id):
     credit = Credit.objects.get(id=id)
+    if check(credit,request): return redirect('home')
     pay = decimal.Decimal(request.POST.get('pay'))
     to_pay = credit.to_pay
     if(pay > to_pay):
@@ -78,9 +85,34 @@ def creditPay(request,id):
     to_pay -= pay
     credit.to_pay = to_pay
     credit.payed += pay
+    
+    if(credit.to_pay == 0.00): 
+        credit.etat = 1
+    else:
+        credit.etat = 0
     credit.save()
     return messages.success(request,"Good")
 
+@login_required
 def creditView(request):
     credits = Credit.objects.filter(user=request.user)
     return render(request, 'credits/credit_view.html',{'credits':credits})
+
+def creditCheck(request):
+    if request.method == 'POST':
+        credit_id = request.POST.get('credit_id')
+        try:
+            credit = Credit.objects.get(id=credit_id)
+            return redirect('credit-detail', id=credit.id)
+        except Credit.DoesNotExist:
+            messages.error(request, 'Credit not found')
+        except ValueError:
+             messages.error(request, 'Invalid Value')
+
+    return render(request, 'credits/credit_find.html')
+
+
+def check(model, request):
+    if(model.user != request.user):
+        messages.error(request, "Can't access this page")
+        return True
